feat: Create the mender-setup CLI tool
This strips out the `setup` functionality from the regular mender-client, and
ports it here, to be a free-standing tool.

In order to do so, I was forced to do a small refac of the mender-client, to
split out the mender-config, to hinder it from pulling in a lot of dependencies.

This ports contains a basic CLI, as well as the `setup.go` file, and a simple
ported `setup_test.go` port.

`deb-requirements` and `support/demo.crt` was added for test compatibility.

Ticket: MEN-6452
Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>